% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_mloglike.R
\name{approx_mloglike}
\alias{approx_mloglike}
\title{Apprimation of the marginal likelihood for btblv models}
\usage{
approx_mloglike(btblv_posterior, N, seed, cores = 1)
}
\arguments{
\item{btblv_posterior}{`btblv_posterior` generated by `btblv::extract_posterior`.}

\item{N}{numeric value with the Monte Carlo sample size.}

\item{seed}{integer value with seed for the random generating system.}

\item{cores}{integer value with the number of cores used for the computation. Default is 1}
}
\value{
object of the class `btblv_approx_mloglike`.
}
\description{
Apprimation of the marginal likelihood for btblv models
}
\examples{
# 
}
