% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_BIC.R
\name{compute_BIC}
\alias{compute_BIC}
\title{Compute the BIC based on the marginal likelihood with respect to latent variables
for the btblv model.}
\usage{
compute_BIC(btblv_posterior, approx_mloglike = NULL, N = NULL, seed, cores = 1)
}
\arguments{
\item{btblv_posterior}{`btblv_posterior` object generated 
by `btblv::extract_posterior`.}

\item{approx_mloglike}{`approx_mloglike` object generated by `tblvArmaUtils::approx_mlog_like`. Default is `NULL`.}

\item{N}{integer value with the Monte Carlo sample size. Default is `NULL`.}

\item{seed}{integer value with seed for the random generating system.}

\item{cores}{number of cores for the computation.}
}
\value{
numeric value with the BIC.
}
\description{
Compute the BIC based on the marginal likelihood with respect to latent variables
for the btblv model.
}
\examples{
# 
}
