% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSE.R
\docType{data}
\name{LSE}
\alias{LSE}
\title{Sum of the log of the exponentials}
\format{
An object of class \code{function} of length 1.
}
\usage{
LSE
}
\arguments{
\item{x}{numeric vector.}
}
\value{
numeric value with sum(log(exp(x)))
}
\description{
Comutes sum(log(exp(x)))
}
\examples{
LSE(c(10, 20, 30))
}
\keyword{datasets}
