% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_mc_log_mlike.R
\docType{data}
\name{rcpp_mc_log_mlike}
\alias{rcpp_mc_log_mlike}
\title{Approximates the marginal likelihood with Monte Carlo integration
for the btblv model.}
\format{
An object of class \code{function} of length 1.
}
\usage{
rcpp_mc_log_mlike
}
\arguments{
\item{N}{integer with Monte Carlo sample size.}

\item{x}{matrix with observed data.}

\item{alpha}{numeric matrix (items x latent dimension size) with the 
estimated alpha.}

\item{beta}{numeric matrix (items x 1) with the estimated beta.}

\item{kappa}{numeric matrix (items x 1) with the estimated kappa.}

\item{phi}{numeric matrix (groups x 1) with the estimated phi.}

\item{sigma}{numeric matrix (groups x 1) with the estimated sigma.}

\item{E_post_sample_list}{List with the latent effect posterior (theta_{t}) 
sample for each time point.}
}
\value{
matrix (iter x time point) with the log of the marginal likelihood 
for each time point. It is still necessary to use `exp()` and `sum()`.
}
\description{
Approximates the marginal likelihood with respect to the latent variables 
with Monte Carlo integration with importance sampling.
}
\examples{
# 
}
\keyword{datasets}
