% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_mc_log_mlike.R
\name{rcpp_mc_log_mlike}
\alias{rcpp_mc_log_mlike}
\title{Approximates the marginal likelihood with Monte Carlo integration
for the btblv model}
\usage{
rcpp_mc_log_mlike(N, x, alpha, beta, kappa, phi, sigma, E_post_sample_list)
}
\arguments{
\item{N}{integer with Monte Carlo sample size.}

\item{x}{matrix with observed data.}

\item{alpha}{numeric matrix (items x latent dimension size) with the 
estimated alpha.}

\item{beta}{numeric matrix (items x 1) with the estimated beta.}

\item{kappa}{numeric matrix (items x 1) with the estimated kappa.}

\item{phi}{numeric matrix (groups x 1) with the estimated phi.}

\item{sigma}{numeric matrix (groups x 1) with the estimated sigma.}

\item{E_post_sample_list}{List with the latent effect posterior 
sample for each time point.}
}
\value{
matrix (iters x 1) with the log of the marginal likelihood sample.
}
\description{
Approximates the marginal likelihood with respect to the latent variables 
with Monte Carlo integration with importance sampling.
}
\examples{
# 
}
